backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
celery[redis]==5.3.6
redis==5.0.1
httpx==0.26.0
aiofiles==23.2.1
Jinja2==3.1.3
markdown==3.5.2
weasyprint==60.2
python-dotenv==1.0.1
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.8.1
groq==0.4.1
beautifulsoup4==4.12.3
lxml==5.1.0
PyYAML==6.0.1
jsonschema==4.21.1
tenacity==8.2.3
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0
python-slugify==8.0.1
pytz==2024.1
python-json-logger==2.0.7
flower==2.0.1
gunicorn==21.2.0
ruff==0.1.15
black==24.1.1
mypy==1.8.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.5.0
